Extract signature utilities
