+++ icpack-js/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "icpack",
+  "version": "1.0.0",
+  "type": "module"
+}
