+++ src/bootstrapper_backend/BootstrapperData.mo
@@ -4,7 +4,7 @@ import Debug "mo:base/Debug";
 import RBTree "mo:base/RBTree";
 import Time "mo:base/Time";
 import Int "mo:base/Int";
-import UserAuth "../lib/UserAuth";
+import UserAuth "mo:icpack-lib/UserAuth";
 import Account "../lib/Account";
 
 persistent actor class BootstrapperData(initialOwner: Principal) = this {
